Move `rates.ts`
