+++ src/data/egg.ts
@@ -22,7 +22,7 @@ import {
   SAME_SPECIES_EGG_SHINY_RATE,
   SHINY_EPIC_CHANCE,
   SHINY_VARIANT_CHANCE,
-} from "#app/data/balance/rates";
+} from "#app/data/rates";
 import { speciesEggTiers } from "#app/data/balance/species-egg-tiers";
 import { speciesStarterCosts } from "#app/data/balance/starters";
 import { pokemonPreEvolutions } from "#app/data/pokemon-pre-evolutions";
